I reverted the AU handing of AUParameterInfo.defaultValue back to using the DfxParam default value since now Bill Stewart says that that in fact is what AU default value is expected to be.
I add stuff to the AU PostConstructor that checks to see if the
initial stream format is actually supported by one of the specified
supported i/o channel-count pairs (if the plugin is one that specifies
those) and, if not, sets the current stream format to what the first
pair specifies (I had another AU failing AUValidation's Render test
because of this problem).

I also figured out that an AU doesn't have to have to be Initialized
to have a valid (and accessible) sampling rate, it only needs to have
had its elements created, which happens in AUBase::PostConstructor
(which calls AUBase::CreateElements).  So now I added a member
variable to DfxPlugin that notices when the elements are created and
now updatesamplerate() checks that member to see if its safe to get
the sampling rate rather than checking if the AU is Initialized.


git-svn-id: https://svn.code.sf.net/p/destroyfx/code/trunk@474 b413c13a-479c-4ba3-9490-20a46969969a