initial import of dfx-au-utilities.* files
These files have some generically useful code for Audio Unit plugins,
which are useful in most plugins or hosts, including ones that I work
on which are not DFX, so that's why this stuff is in separate files.
Also, in order to be able to use this source code in any other
project, it is offered as public domain rather than under GPL like our
other source code.


git-svn-id: https://svn.code.sf.net/p/destroyfx/code/trunk@487 b413c13a-479c-4ba3-9490-20a46969969a